refactor: Lotto 클래스 package 변경(lotto.domain.entity)
